Merge simplification w/ everything else
Previously simplification had its own separate logic because I did not
think it would be possible with 'rules'.

Now it is implemented with rules which is both pretty cool and also
practical, because I don't have to worry about a separate task of
passing the output of a derivative to a simplifier.